Move app specific analytics logic to a separate "processing" function. This separates concerns and makes testing easier
