+++ packages/lib/src/core/Analytics/Analytics.ts
@@ -1,24 +1,12 @@
 import LogEvent from '../Services/analytics/log-event';
 import CollectId from '../Services/analytics/collect-id';
 import EventsQueue, { EventsQueueModule } from './EventsQueue';
-import { ANALYTICS_EVENT, AnalyticsInitialEvent, AnalyticsObject, AnalyticsProps, CreateAnalyticsEventObject, SendAnalyticsObject } from './types';
-import {
-    ANALYTICS_EVENT_ERROR,
-    ANALYTICS_EVENT_INFO,
-    ANALYTICS_EVENT_LOG,
-    ANALYTICS_FOCUS_STR,
-    ANALYTICS_INFO_TIMER_INTERVAL,
-    ANALYTICS_PATH,
-    ANALYTICS_RENDERED_STR,
-    ANALYTICS_SELECTED_STR,
-    ANALYTICS_SUBMIT_STR,
-    ANALYTICS_UNFOCUS_STR,
-    ANALYTICS_VALIDATION_ERROR_STR
-} from './constants';
+import { ANALYTICS_EVENT, AnalyticsInitialEvent, AnalyticsObject, AnalyticsProps, CreateAnalyticsEventObject } from './types';
+import { ANALYTICS_EVENT_ERROR, ANALYTICS_EVENT_INFO, ANALYTICS_EVENT_LOG, ANALYTICS_INFO_TIMER_INTERVAL, ANALYTICS_PATH } from './constants';
 import { debounce } from '../../components/internal/Address/utils';
 import { AnalyticsModule } from '../../components/types';
 import { createAnalyticsObject } from './utils';
-import { THREEDS2_ERROR, THREEDS2_FULL } from '../../components/ThreeDS2/config';
+import { analyticsPreProcessor } from './analyticsPreProcessor';
 
 let capturedCheckoutAttemptId = null;
 let hasLoggedPixel = false;
@@ -124,80 +112,10 @@ const Analytics = ({ loadingContext, locale, clientKey, analytics, amount, analy
 
         getEnabled: () => props.enabled,
 
-        sendAnalytics: (component: string, analyticsObj: SendAnalyticsObject) => {
-            const { type, target } = analyticsObj;
-
-            switch (type) {
-                // Called from BaseElement (when component mounted) or, from DropinComponent (after mounting, when it has finished resolving all the PM promises)
-                // &/or, from DropinComponent when a PM is selected
-                case ANALYTICS_RENDERED_STR: {
-                    const { isStoredPaymentMethod, brand } = analyticsObj;
-                    const data = { component, type, isStoredPaymentMethod, brand };
-
-                    // AnalyticsAction: action: 'event' type:'rendered'|'selected'
-                    anlModule.createAnalyticsEvent({
-                        event: 'info',
-                        data
-                    });
-                    break;
-                }
-
-                case ANALYTICS_SUBMIT_STR:
-                    // PM pay button pressed - AnalyticsAction: action: 'log' type:'submit'
-                    anlModule.createAnalyticsEvent({
-                        event: 'log',
-                        data: { component, type, target: 'pay_button', message: 'Shopper clicked pay' }
-                    });
-                    break;
-
-                case ANALYTICS_FOCUS_STR:
-                case ANALYTICS_UNFOCUS_STR:
-                    anlModule.createAnalyticsEvent({
-                        event: 'info',
-                        data: { component, type, target }
-                    });
-                    break;
-
-                case ANALYTICS_VALIDATION_ERROR_STR: {
-                    const { validationErrorCode, validationErrorMessage } = analyticsObj;
-                    anlModule.createAnalyticsEvent({
-                        event: 'info',
-                        data: { component, type, target, validationErrorCode, validationErrorMessage }
-                    });
-                    break;
-                }
-
-                case ANALYTICS_SELECTED_STR:
-                    anlModule.createAnalyticsEvent({
-                        event: 'info',
-                        data: { component, type, target }
-                    });
-                    break;
-
-                case THREEDS2_FULL: {
-                    const { message } = analyticsObj;
-                    anlModule.createAnalyticsEvent({
-                        event: 'log',
-                        data: { component, type, message }
-                    });
-                    break;
-                }
-
-                case THREEDS2_ERROR: {
-                    const { message, code, errorType } = analyticsObj;
-                    anlModule.createAnalyticsEvent({
-                        event: 'error',
-                        data: { component, type, message, code, errorType }
-                    });
-                    break;
-                }
-
-                default: {
-                    anlModule.createAnalyticsEvent(analyticsObj as CreateAnalyticsEventObject);
-                }
-            }
-        }
+        sendAnalytics: null
     };
+
+    anlModule.sendAnalytics = analyticsPreProcessor(anlModule);
 
     return anlModule;
 };
